refactor: use router link in snackbar
