+++ src/App.tsx
@@ -1,10 +1,20 @@
 import { createTheme, CssBaseline, Grow, ThemeProvider, useMediaQuery } from '@mui/material';
 import { SnackbarProvider } from 'notistack';
-import { useMemo } from 'react';
-import AppRouter from './AppRouter';
+import { lazy, Suspense, useMemo, useState } from 'react';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import AppErrorBoundary from './components/AppErrorBoundary';
+import Layout from './components/Layout';
+import Loading from './components/Loading';
 import SnackbarDismissButton from './components/Snackbars/SnackbarDismissButton';
+import Home from './pages/Home';
+import NotFound from './pages/NotFound';
+
+const Post = lazy(() => import('./pages/Post'));
+const Update = lazy(() => import('./pages/Update'));
 
 function App() {
+	const [isLoading, setIsLoading] = useState(false);
+
 	const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)');
 
 	const theme = useMemo(
@@ -42,13 +52,42 @@ function App() {
 		<>
 			<ThemeProvider theme={theme}>
 				<CssBaseline />
-				<SnackbarProvider
-					maxSnack={5}
-					TransitionComponent={Grow}
-					action={(snackbarKey) => <SnackbarDismissButton snackbarKey={snackbarKey} />}
-				>
-					<AppRouter />
-				</SnackbarProvider>
+				<Loading isLoading={isLoading} />
+				<BrowserRouter>
+					<SnackbarProvider
+						maxSnack={5}
+						TransitionComponent={Grow}
+						action={(snackbarKey) => <SnackbarDismissButton snackbarKey={snackbarKey} />}
+					>
+						<Routes>
+							<Route path="*" element={<Layout />}>
+								<Route path="" element={<Home />} />
+								<Route path="home" element={<Home />} />
+								<Route
+									path="post"
+									element={
+										<AppErrorBoundary>
+											<Suspense fallback={<Loading isLoading={true} />}>
+												<Post setIsLoading={setIsLoading} />
+											</Suspense>
+										</AppErrorBoundary>
+									}
+								/>
+								<Route
+									path="update"
+									element={
+										<AppErrorBoundary>
+											<Suspense fallback={<Loading isLoading={true} />}>
+												<Update setIsLoading={setIsLoading} />
+											</Suspense>
+										</AppErrorBoundary>
+									}
+								/>
+								<Route path="*" element={<NotFound />} />
+							</Route>
+						</Routes>
+					</SnackbarProvider>
+				</BrowserRouter>
 			</ThemeProvider>
 		</>
 	);
