+++ src/TypeAugments.d.ts
@@ -1,5 +1,4 @@
 import type { Theme } from '@mui/material';
-import type { OptionsObject, SnackbarKey, SnackbarMessage } from 'notistack';
 
 declare module 'react' {
 	interface HTMLAttributes<T> {
@@ -9,9 +8,4 @@ declare module 'react' {
 
 declare module '@mui/styles/defaultTheme' {
 	interface DefaultTheme extends Theme {}
-}
-
-declare global {
-	function enqueueSnackbar(message: SnackbarMessage, options?: OptionsObject | undefined): SnackbarKey;
-	function closeSnackbar(key?: SnackbarKey): void;
 }
